Revised p-value computation for greater numerical stability
With the use of the logarithm of the gamma function, Hodges' exact formula for the p-value may be used reliably for sample sizes up to at least 10^7, if not greater. For n < 50, where the exact formula was previously implemented using nchoosek, the implementation using gammaln to compute the binomial coefficients agrees with that using nchoosek.